requirements.txt
fastapi
uvicorn
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv